requirements.txt
Django==4.1.7
pytz==2022.7.1
djongo==1.3.6
pymongo==3.12.1
dnspython
djangorestframework==3.14.0
asgiref==3.6.0
autopep8==2.0.2
pycodestyle==2.10.0
tzdata==2022.7
djangorestframework-simplejwt
django-cors-headers